requirements-curator.txt
watchdog>=3.0.0
PyGithub>=2.3.0
PyYAML>=6.0.2
markdown-it-py>=3.0.0
